refactor: :recycle: 멤버쉽 서비스 테스트 리팩토링
