+++ rest/http-client.private.env.json
@@ -1,0 +1,5 @@
+{
+  "dev": {
+    "endpoint": "http://localhost:8080"
+  }
+}
