--- server/rest/http-client.private.env.json
@@ -1,5 +1,0 @@
-{
-  "dev": {
-    "endpoint": "http://localhost:8080"
-  }
-}
