+++ tmp/build-errors.log
@@ -1,1 +1,1 @@
-exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 2exit status 1
+exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 1exit status 2exit status 1exit status 1exit status 1
